requirements.txt
playwright>=1.40
pydantic>=2.0
pyyaml>=6.0
rich>=13.0
click>=8.0
beautifulsoup4>=4.12
lxml>=5.0
Pillow>=10.0
